Fix Directory Number Collisions Check
This commit fixes the previous check would not allow for additional source files to be added to an existing problem folder. Now the program correctly allows target directories that are identical with an existing problem directory (such as when adding additional source files) and rejects new target directories if the number collides with an existing problem directory.